+++ src/lib.rs
@@ -2,7 +2,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! easy-ext = "0.2"
+//! easy-ext = "1"
 //! ```
 //!
 //! *Compiler support: requires rustc 1.31+*