+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [0.2.8] - 2021-06-23
 
+**Note: This release has been yanked because of regression which fixed in 0.2.9.**
+
 - [Support specifying visibility directly on `impl`.](https://github.com/taiki-e/easy-ext/pull/31)
 
   ```rust